[iOS] Fix remaining misuses of abs() and fabsf() https://bugs.webkit.org/show_bug.cgi?id=130947
Reviewed by Simon Fraser.

Fixed issues found when compiling with -Wabsolute-value.

* platform/ios/ScrollAnimatorIOS.mm:
(WebCore::ScrollAnimatorIOS::handleTouchEvent): Called abs() instead of fabsf for integral types. Explicitly
cast the divisor to float in the call to atanf().
* platform/ios/wak/WKView.mm:
(_WKViewAutoresizeCoord): Explicitly cast origMarginsTotal to an int before calling abs(). Added a FIXME
indicating the strange behavior when origMarginsTotal is in the range (0, 1) (as pointed out by Darin).

Canonical link: https://commits.webkit.org/155297@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@174383 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/platform/ios/wak/WKView.mm
@@ -247,8 +247,9 @@ static void _WKViewAutoresizeCoord(bool bByHeight, unsigned int sizingMethod, co
             }
                 // Do the 50/50 split even if XorY = 0 and the WidthOrHeight is only
                 // one pixel shorter...
+                // FIXME: If origMarginsTotal is in the range (0, 1) then we won't do the 50/50 split. Is this right?
                 else if (origMarginsTotal == 0.0f 
-                         || (abs(origMarginsTotal) == 1)) {
+                    || (abs(static_cast<int>(origMarginsTotal)) == 1)) {
                     prop = 0.5f;  // Then split it 50:50.
                 }
                 else {
